requirements.txt
Django
djangorestframework
django-cors-headers
django-channels
channels-redis
web3
py-solc-x
python-dotenv
cryptography
celery
redis
pytest
pytest-django
requests
rest_framework_simplejwt